Split user view tests into separate test files for multiple endpoints
